+++ inst/include/simmer/common.h
@@ -1,5 +1,5 @@
 // Copyright (C) 2015-2016 Bart Smeets and Iñaki Ucar
-// Copyright (C) 2016-2018 Iñaki Ucar
+// Copyright (C) 2016-2018,2020 Iñaki Ucar
 //
 // This file is part of simmer.
 //
@@ -44,10 +44,12 @@
 #include <boost/optional.hpp>
 #include <boost/typeof/typeof.hpp>
 #include <boost/any.hpp>
+#include <boost/shared_ptr.hpp>
 
 #define OPT   boost::optional
 #define NONE  boost::none
 #define ANY   boost::any
+#define SHD   boost::shared_ptr
 
 #include <boost/foreach.hpp>